1. finalized all prediction and initial training parts 2. added models and data to github

+++ main_service/models.py
@@ -7,7 +7,7 @@ class ModelResult(models.Model):
     day_num = models.IntegerField(default=0)
     ema_order = models.SmallIntegerField(default=0)
     prediction_result = models.SmallIntegerField(default=-1)
-    accuracy = models.SmallIntegerField(default=0)
+    accuracy = models.FloatField(default=0)
     feature_ids = models.TextField()
     model_tag = models.BooleanField(default=False)
     user_tag = models.BooleanField(default=False)
